+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- Private message encryption/decryption with authentification
+
 ## [0.12.1] - 2020-03-03
 
 ### Fixed